# app/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go - RESTful",
        "title": "Go Rest API Starter Doc",
        "termsOfService": "https://swagger.io/terms/",
        "contact": {
            "name": "GulizAY",
            "url": "https://github.com/gulizay91",
            "email": "gulizay91@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8091",
    "basePath": "/",
    "paths": {
        "/api/v1/user": {
            "post": {
                "description": "create user",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "create user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.UserModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{id}": {
            "delete": {
                "description": "delete user",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "delete user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/user/{subId}": {
            "get": {
                "description": "get user",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subId",
                        "name": "subId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{subId}/media": {
            "get": {
                "description": "get user's images",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subId",
                        "name": "subId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    }
                }
            }
        },
        "/api/v1/user/{subId}/upload-media": {
            "patch": {
                "description": "upload user's images",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "upload images",
                "parameters": [
                    {
                        "type": "string",
                        "description": "subId",
                        "name": "subId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "Show the status of server.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_gulizay91_go-rest-api_pkg_models.Gender": {
            "type": "string",
            "enum": [
                "female",
                "male",
                "non-binary",
                "intersex"
            ],
            "x-enum-varnames": [
                "Female",
                "Male",
                "NonBinary",
                "Intersex"
            ]
        },
        "github_com_gulizay91_go-rest-api_pkg_models.Language": {
            "type": "string",
            "enum": [
                "tr",
                "eng"
            ],
            "x-enum-varnames": [
                "TR",
                "ENG"
            ]
        },
        "github_com_gulizay91_go-rest-api_pkg_models.Media": {
            "type": "object",
            "required": [
                "images"
            ],
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_gulizay91_go-rest-api_pkg_models.ServiceResponseModel": {
            "type": "object",
            "properties": {
                "data": {},
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "github_com_gulizay91_go-rest-api_pkg_models.UserModel": {
            "type": "object",
            "required": [
                "birthDate",
                "email",
                "firstName",
                "lastName",
                "subId"
            ],
            "properties": {
                "birthDate": {
                    "type": "string"
                },
                "createdDate": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string",
                    "maxLength": 50
                },
                "gender": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.Gender"
                        }
                    ],
                    "example": "female"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.Language"
                    }
                },
                "lastName": {
                    "type": "string",
                    "maxLength": 50
                },
                "media": {
                    "$ref": "#/definitions/github_com_gulizay91_go-rest-api_pkg_models.Media"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "subId": {
                    "type": "string"
                },
                "updatedDate": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Bearer-Token",
            "in": "header"
        }
    }
}